commity

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -3,13 +3,13 @@
 /**
  * print_listint - fu
  * @h: in
- * Description: none
+ *
  * Return: count
  */
 
 size_t print_listint(const listint_t *h)
 {
-	int count = 0;
+	size_t count = 0;
 
 	while (h)
 	{
